remove unused defines, join caching code with esp vfs code to reduce exported symbols and shrink final file size
