+++ app/sqlite3/Makefile
@@ -15,7 +15,7 @@ ifndef PDIR
 GEN_LIBS = libsqlite3.a
 endif
 
-STD_CFLAGS=-std=gnu11 -Wimplicit -Wno-undef -I../platform -DENABLE_LOCAL_MEMDEBUG
+STD_CFLAGS=  -std=gnu11 -Wimplicit -Wno-undef
 STD_CFLAGS+= -DSQLITE_CORE=1
 STD_CFLAGS+= -DSQLITE_NO_SYNC=1
 STD_CFLAGS+= -DYYSTACKDEPTH=20
@@ -31,7 +31,7 @@ STD_CFLAGS+= -DSQLITE_DISABLE_LFS=1
 STD_CFLAGS+= -DSQLITE_DISABLE_DIRSYNC=1
 STD_CFLAGS+= -DSQLITE_DISABLE_FTS3_UNICODE=1
 STD_CFLAGS+= -DSQLITE_DISABLE_FTS4_DEFERRED=1
-STD_CFLAGS+= -DSQLITE_LIKE_DOESNT_MATCH_BLOBS
+STD_CFLAGS+= -DSQLITE_LIKE_DOESNT_MATCH_BLOBS=1
 STD_CFLAGS+= -DSQLITE_DEFAULT_CACHE_SIZE=-1
 STD_CFLAGS+= -DSQLITE_DEFAULT_MEMSTATUS=0
 STD_CFLAGS+= -DSQLITE_DEFAULT_MMAP_SIZE=0
@@ -89,7 +89,6 @@ STD_CFLAGS+= -DSQLITE_OMIT_VIEW=1
 STD_CFLAGS+= -DSQLITE_OMIT_VIRTUALTABLE=1
 STD_CFLAGS+= -DSQLITE_OMIT_WAL=1
 STD_CFLAGS+= -DSQLITE_OMIT_XFER_OPT=1
-STD_CFLAGS+= -DSQLITE_OMIT_XFER_OPT=1
 #STD_CFLAGS+= -DSQLITE_OMIT_COMPLETE=1
 #STD_CFLAGS+= -DSQLITE_OMIT_SUBQUERY=1
 #STD_CFLAGS+= -DSQLITE_OMIT_DATETIME_FUNCS=1
@@ -119,5 +118,6 @@ STD_CFLAGS+= -DSQLITE_OMIT_XFER_OPT=1
 INCLUDES := $(INCLUDES) -I $(PDIR)include
 INCLUDES += -I ./
 INCLUDES += -I ../libc
+INCLUDES += -I ../platform
 PDIR := ../$(PDIR)
 sinclude $(PDIR)Makefile
