--- app/sqlite3/caching.c
@@ -1,146 +1,0 @@
-#include <c_stdlib.h>
-#include <c_string.h>
-#include <c_stdint.h>
-#include <sqlite3.h>
-#include <caching.h>
-
-static uint32_t linkedlist_store (linkedlist_t **leaf, uint32_t offset, uint32_t len, const uint8_t *data) {
-	const uint8_t blank[CACHEBLOCKSZ] = { 0 };
-	uint16_t blockid = offset/CACHEBLOCKSZ;
-	linkedlist_t *block;
-
-	if (!memcmp(data, blank, CACHEBLOCKSZ))
-		return len;
-
-	block = *leaf;
-	if (!block || ( block->blockid != blockid ) ) {
-		block = sqlite3_malloc ( sizeof( linkedlist_t ) );
-		if (!block)
-			return SQLITE_NOMEM;
-
-		memset (block->data, 0, CACHEBLOCKSZ);
-		block->blockid = blockid;
-	}
-
-	if (!*leaf) {
-		*leaf = block;
-		block->next = NULL;
-	} else if (block != *leaf) {
-		if (block->blockid > (*leaf)->blockid) {
-			block->next = (*leaf)->next;
-			(*leaf)->next = block;
-		} else {
-			block->next = (*leaf);
-			(*leaf) = block;
-		}
-	}
-
-	memcpy (block->data + offset%CACHEBLOCKSZ, data, len);
-
-	return len;
-}
-
-uint32_t filecache_pull (pFileCache_t cache, uint32_t offset, uint32_t len, uint8_t *data) {
-	uint16_t i;
-	float blocks;
-	uint32_t r = 0;
-
-	blocks = ( offset % CACHEBLOCKSZ + len ) / (float) CACHEBLOCKSZ;
-
-	if (blocks == 0.0)
-		return 0;
-
-	if (( blocks - (int) blocks) > 0.0)
-		blocks = blocks + 1.0;
-
-	for (i = 0; i < (uint16_t) blocks; i++) {
-		uint16_t round;
-		float relablock;
-		linkedlist_t *leaf;
-		uint32_t relaoffset, relalen;
-		uint8_t * reladata = (uint8_t*) data;
-
-		relalen = len - r;
-
-		reladata = reladata + r;
-		relaoffset = offset + r;
-
-		round = CACHEBLOCKSZ - relaoffset%CACHEBLOCKSZ;
-		if (relalen > round) relalen = round;
-
-		for (leaf = cache->list; leaf && leaf->next; leaf = leaf->next) {
-			if ( ( leaf->next->blockid * CACHEBLOCKSZ ) > relaoffset )
-				break;
-		}
-
-		relablock = relaoffset/((float)CACHEBLOCKSZ) - leaf->blockid;
-
-		if ( ( relablock >= 0 ) && ( relablock < 1 ) )
-			memcpy (data + r, leaf->data + (relaoffset % CACHEBLOCKSZ), relalen);
-
-		r = r + relalen;
-	}
-
-	return 0;
-}
-
-uint32_t filecache_push (pFileCache_t cache, uint32_t offset, uint32_t len, const uint8_t *data) {
-	uint16_t i;
-	float blocks;
-	uint32_t r = 0;
-	uint8_t updateroot = 0x1;
-
-	blocks = ( offset % CACHEBLOCKSZ + len ) / (float) CACHEBLOCKSZ;
-
-	if (blocks == 0.0)
-		return 0;
-
-	if (( blocks - (int) blocks) > 0.0)
-		blocks = blocks + 1.0;
-
-	for (i = 0; i < (uint16_t) blocks; i++) {
-		uint16_t round;
-		uint32_t localr;
-		linkedlist_t *leaf;
-		uint32_t relaoffset, relalen;
-		uint8_t * reladata = (uint8_t*) data;
-
-		relalen = len - r;
-
-		reladata = reladata + r;
-		relaoffset = offset + r;
-
-		round = CACHEBLOCKSZ - relaoffset%CACHEBLOCKSZ;
-		if (relalen > round) relalen = round;
-
-		for (leaf = cache->list; leaf && leaf->next; leaf = leaf->next) {
-			if ( ( leaf->next->blockid * CACHEBLOCKSZ ) > relaoffset )
-				break;
-			updateroot = 0x0;
-		}
-
-		localr = linkedlist_store(&leaf, relaoffset, (relalen > CACHEBLOCKSZ) ? CACHEBLOCKSZ : relalen, reladata);
-		if (localr == SQLITE_NOMEM)
-			return SQLITE_NOMEM;
-
-		r = r + localr;
-
-		if (updateroot & 0x1)
-			cache->list = leaf;
-	}
-
-	if (offset + len > cache->size)
-		cache->size = offset + len;
-
-	return r;
-}
-
-void filecache_free (pFileCache_t cache) {
-	pLinkedList_t this = cache->list, next;
-
-	while (this != NULL) {
-		next = this->next;
-		sqlite3_free (this);
-		this = next;
-	}
-}
